+++ src/gufo/err/abc/__init__.py
@@ -3,4 +3,13 @@
 # ---------------------------------------------------------------------
 # Copyright (C) 2022-23, Gufo Labs
 # ---------------------------------------------------------------------
-"""Abstract base classes to be used to extend the functionality."""
+"""
+Abstract base classes to be used to extend the functionality.
+
+* [FailFast][gufo.err.abc.failfast.BaseFailFast]:
+  Fail-fast processing.
+* [Formatter][gufo.err.abc.formatter.BaseFormatter]:
+  Traceback formatters.
+* [Middleware][gufo.err.abc.middleware.BaseMiddleware]:
+  Error-processing middleware.
+"""