+++ CHANGELOG.md
@@ -10,6 +10,8 @@ To see unreleased changes, please see the [CHANGELOG on the master branch](https
 
 ### Added
 
+* `BaseFormatter` abstract base class.
+* `TerseFormatter` and `ExtendFormatter` formatters.
 * `ErrorInfo` got optional `root_module` field.
 * `ErrorInfo.get_top_app_frame()` function.
 * `py.typed` file for PEP-561 compatibility